Add polygon and bbox attributes to PageNode
This is intended to simplify serialization of documents without nested
regions. Both PAGE and ALTO wants to put text in regions. If HTRFLOW
puts text directly on the input image (and not in a region), the parent
PageNode is treated as the parent "region", and thus needs a bounding
box and polygon.