Clarify Volume architecture
PageNodes with text but without any regions needed special treatment and
caused many hacks of varying degree of complexity (see for example the
old ALTO template). With this commit, PageNodes cannot contain any text.
If a PageNode is assigned text, it spawns a new child and assigns the
text to it. This will hopefully simplify typechecking and handling of
volumes.