Update serialization
Here comes a major update of the serialization module

- PAGE output allows nested regions (unlike ALTO), but the previous
implementation flattened the tree to the same format as ALTO (max. one
level of segmentation). With this update, Page files keep the original
nested segmentation.
- All output formats have gotten their own classes. This makes it easier
to add more formats if needed, and at the same time, the jinja templates
(for the formats that use them) are only loaded once.
- The jinja templates are simplified and their variable names updated to
match the python code.