Clean volume.py and update tests
Tried to solve some mypy complaints in volume.py, which resulted in
splitting the Node class into two: One base class for purely tree-
related stuff (traversal, printing, etc) and a extender class which acts
as a shared interface for RegionNode and PageNode.

Updated the tests accordingly and fixed a bug in the serialization tests.