Unify result classes
The previous implementation distinguished between results from
segmentation and text recognition. There may however be cases where
we want elements from both, for example in segmenting already
recognized text lines into words. Instead of creating a combination
class for those situations, this commit gets rid of the separation
and uses the generic `Result` class for all cases.

This means that Volume doesn't need to distinguish between different
types of Result. See Volume.update for the updated (:)) version that
handles the generic Result. It also means that the interface for
creating results is slightly different, and that the models need to be
updated.