Convert attributes to properties
Some mostly cosmetic updates to the volume module.
- Node properties related to geometry (bounding boxes, polygons,
coordinates) are converted into properties. This makes them read-only,
which is good, because we don't want to make any changes to the results
once they are registered (.update()'d) in the volume. This also makes it
easier to document them.
- Moved some methods from RegionNode to BaseRegionNode, so that they are
available for PageNodes as well.
- Wrote some docstrings