Add Polygon and Point, rewrite Segment's __init__
Geometry module:
- Added Polygon and Point classes
- Updated the Bbox class
- Replaced most of the geometric calculations in other parts of HTRFLOW
with the new additions in this module.

Results module:
- Removed the specific .from_bbox and .from_mask constructors, as these
are redundant since a previous change to Segment's init.
- Segments can now be initialized from a mask, bbox or polygon (new!)
or any combination of the three. Tried to make the quite involved logic
cleaner by using a match/case statement, but still not sure if it is the
cleanest option.
- Anyways, this change uses many of the new functions from the geometry
module, which is why these two come bundled together.