Remove ability to export unsegmented pages to XML
This commit removes some quirks and complexity in volume.py that were
added to enable XML export of pages with text but without any regions.
Instead an error is raised whenever the user tries to export such pages
to XML. JSON and plaintext is still available for those cases (and also
a much more suitable option). If there ever is a need for handling these
cases, it should be handled in the serialiation module, and not by
modifying Volume's internals.