Classification models preparations
- Result now takes two types of data: (1) segments and (2) anything
else. (2) could be text recognition results or classification results,
or other kinds of data. Volume will as usual segment the leaf nodes if
given a segmentation result, while data of type (2) will be added to
the leaf nodes' data dictionaries.
- Added a dummy classification model to use as a template for real
classification models. It represents its result as a dictionary; we
might want to implement a dataclass instead, analogous to
RecgonizedText.
- Changed some tests to reflect the new results interface.