Update PAGE template
Old template produced a large TextRegion that covered the entire page.
It was needed when the page only contained TextLines (which must be put
in a TextRegion), but it was rendered elsewise too. The transkribus
client was not able to parse these files. Now the big region should only
be added when needed.