Update internals of volume
Here comes a little backstory of Volume and Nodes. At first, nodes kept
their data in attributes. This proved to be a bit too inflexible, since
the class needed modification every time new kinds of data were added
to the nodes. (Latest example of this is the implementation of the
classification model).

The solution to that problem was to give the Node base class a `data`
attribute, which is a dicitionary, where all kinds of data/fields/
attributes could be stored. This solution also made it super easy to
export everything to json. This also made the PageNode/RegionNode
classes a lot lighter.

However, that solution led to EVERYTHING being exported to json, even
data that is only used within HTRFLOW and doesn't need exporting. So,
this commit makes Volume/Nodes store data in a hybrid way:

- All attributes that are derived from the segmentation results to use
within HTRFLOW (for example, global/local coordinates, global/local
polygons) are stored as Node attributes. These things are accessed
during runtime and thus motivates a proper attribute.
- All other data (typically results from models) is stored in the data
dictionary. This is the stuff that is exported to JSON.