Backup volume between pipeline steps
- pipeline pickles the volume between each step
- the auto_import function now accepts a path to a pickled volume
- pipeline.run() takes a `start` argument which specifies which
step to start from
- the start argument may be passed from the command line arguments
to main.py (where it is called --resume)

In total, this makes it possible to restart pipeline execution if the
pipeline failed, by passing the pickled volume as input and specifying
the resume step:

python main.py my-pipeline.yaml .cache/volume.pickle --resume 2

This reruns step 2 and continues from there.

Still left todo: Optimize the size of the pickles, perhaps by not
including the images.