Change reading order of regions
Turns out left-right-top-down ordering of regions is not a good enough
strategy. Instead, a similar approach as the one used for the lines
within each region is used: Regions are sorted according to their
location relative to the page's printspace, and then top-down. There is
also an option to sort according to a two-page layout.