Change handling of line- and word nodes
With this change, a Result instance can hold text recognition
results on both word- and line level. (Previously only one of
them could be handled at a time). This allows WordlevelTrocr
(and other word-level models or word segmentation methods) to
report confidence scores on both word- and line level. This is
needed in preparation for AB#608.